updated tutorial notebooks to match new parser module. Moved all parsers to seperate module.
