--- snorkel/contrib/brat/__init__.py
@@ -1,2 +1,0 @@
-from tools import Brat
-
