--- snorkel/contrib/brat/tmpl.config
@@ -1,18 +1,0 @@
-# -*- Mode: Text; tab-width: 2; indent-tabs-mode: nil; coding: utf-8; -*-
-# vim:set ft=conf ts=2 sw=2 sts=2 autoindent:
-
-[entities]
-
-{}
-
-[relations]
-
-{}
-
-[events]
-
-{}
-
-[attributes]
-
-{}
