+++ snorkel/parsers/__init__.py
@@ -1,0 +1,4 @@
+from .parser import *
+from .doc_preprocessors import *
+from .corenlp import *
+from .spacy import *
